+++ bin/otrs.FillDB.pl
@@ -170,7 +170,7 @@ EOF
             Title        => RandomSubject(),
             QueueID      => $QueueIDs[ int( rand($#QueueIDs) ) ],
             Lock         => 'unlock',
-            Priority     => '3 normal',
+            Priority     => PriorityGet(),
             State        => 'new',
             CustomerNo   => int( rand(1000) ),
             CustomerUser => RandomAddress(),
@@ -497,6 +497,18 @@ sub RandomBody {
         $Body .= $Text[ int( rand( $#Text + 1 ) ) ] . "\n";
     }
     return $Body;
+}
+
+sub PriorityGet {
+    my %PriorityList = $Kernel::OM->Get('Kernel::System::Priority')->PriorityList(
+        Valid => 1,
+    );
+
+    # possible not make this one sorted - for more variation in ticket-priority.
+    for my $Prio ( sort keys %PriorityList ) {
+        return $PriorityList{$Prio};
+    }
+    return '3 normal';
 }
 
 sub QueueGet {